Backend maintenance.
- Extract utility functions to a separate file. This functions are related to the configuration files.
- Delete the admin mode. It is no longer needed due to RBAC.
