+++ SplunkAppForWazuh/appserver/static/js/config/routes/settings-states.js
@@ -27,16 +27,6 @@ define(['../module'], function(module) {
                   return false;
                 }
               }
-            ],
-            isAdmin: [
-              '$currentDataService',
-              async $currentDataService => {
-                try {
-                  return await $currentDataService.isAdmin()
-                } catch (error) {
-                  return false
-                }
-              }
             ]
           }
         })
