+++ SplunkAppForWazuh/appserver/static/js/controllers/settings/main/settingsCtrl.js
@@ -2,12 +2,11 @@ define(['../../module'], function(controllers) {
   'use strict'
 
   class Main {
-    constructor($scope, $navigationService, isAdmin, isWazuhAdmin) {
+    constructor($scope, $navigationService, isWazuhAdmin) {
       this.navigationService = $navigationService
       this.scope = $scope
       this.scope.message = 'Settings'
       this.scope.tabName = ''
-      this.scope.isAdmin = isAdmin
       this.scope.isWazuhAdmin = isWazuhAdmin
     }
