+++ SplunkAppForWazuh/bin/utils/__init__.py
@@ -1,3 +1,4 @@
 # __init__.py
 from .dict import *
 from .params import *
+from .config import *
