+++ SplunkAppForWazuh/default/config.conf
@@ -15,6 +15,5 @@ nist = false
 reporting = true
 
 [configuration]
-admin = true
 log.level = info
 timeout = 20
